feat: enforce consensus settings (#851)
## Purpose of Changes and their Description

Define a bunch of recommended settings for both `app.toml` and
`config.toml` and enforce them upon start-up.

These settings are automatically set as default when using the `init`
command, and if changed they are overridden before loading the
application.

In order to still allow their modification a
`--unsafe-override-recommended-settings` flag has been added to the
`start` command to bypass the enforcing logic and let the values
provided in the `config.toml` and `app.toml` files.

The recommended settings to enforce that are set through this PR are the
following:
- `consensus.timeout_propose`: 5s
- `consensus.timeout_propose_delta`: 500ms
- `consensus.timeout_prevote`: 3s
- `consensus.timeout_prevote_delta`:  500ms
- `consensus.timeout_precommit`: 3s
- `consensus.timeout_precommit_delta`: 500ms
- `consensus.timeout_commit`: 5s

## Are these changes tested and documented?

- [x] If tested, please describe how. If not, why tests are not needed.
- [x] If documented, please describe where. If not, describe why docs
are not needed.
- [x] Added to `Unreleased` section of `CHANGELOG.md`?